+++ src/colors/palette/defaults/DefaultPalettes.h
@@ -13,26 +13,26 @@ namespace lw::colors::palettes
 {
 template <typename TColor, typename = std::enable_if_t<ColorType<TColor>>> class StaticPalette : public IPalette<TColor>
 {
-    public:
-        using StopsView = typename IPalette<TColor>::StopsView;
+  public:
+    using StopsView = typename IPalette<TColor>::StopsView;
 
-        StaticPalette() = default;
+    StaticPalette() = default;
 
-        // Caller convention: stops are expected in non-decreasing index order.
-        // Duplicate indexes are allowed and create zero-width transitions.
-        explicit StaticPalette(StopsView stops) : _stops(stops) {}
+    // Caller convention: stops are expected in non-decreasing index order.
+    // Duplicate indexes are allowed and create zero-width transitions.
+    explicit StaticPalette(StopsView stops) : _stops(stops) {}
 
-        template <size_t N>
-        explicit StaticPalette(const std::array<PaletteStop<TColor>, N>& stops) : _stops(stops.data(), stops.size())
-        {
-        }
+    template <size_t N>
+    explicit StaticPalette(const std::array<PaletteStop<TColor>, N>& stops) : _stops(stops.data(), stops.size())
+    {
+    }
 
-        StopsView stops() const override { return _stops; }
+    StopsView stops() const override { return _stops; }
 
-        void update(uint8_t = 0) override {}
+    void update(uint32_t = 0) override {}
 
-    private:
-        StopsView _stops{};
+  private:
+    StopsView _stops{};
 };
 } // namespace lw::colors::palettes
 
@@ -41,7 +41,7 @@ namespace lw::palettes
 template <typename TColor = colors::DefaultColorType> struct NamedPalette
 {
     const char* name;
-        colors::palettes::StaticPalette<TColor> (*create)();
+    colors::palettes::StaticPalette<TColor> (*create)();
 };
 
 namespace default_palette_data
